# docs/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "TaxiCity API",
		"description": "Comprehensive REST API for the TaxiCity taxi booking platform",
		"version": "1.0.0",
		"contact": {
			"name": "TaxiCity Support",
			"email": "support@taxiciti.com"
		},
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000/api",
			"description": "Development server"
		},
		{
			"url": "https://taxiciti.vercel.app/api",
			"description": "Production server"
		}
	],
	"security": [
		{
			"ClerkAuth": []
		}
	],
	"components": {
		"securitySchemes": {
			"ClerkAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Clerk JWT authentication token"
			}
		},
		"schemas": {
...
			"Rank": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"location": {
						"type": "object",
						"properties": {
							"latitude": {
								"type": "number",
								"minimum": -90,
								"maximum": 90
							},
							"longitude": {
								"type": "number",
								"minimum": -180,
								"maximum": 180
							}
						}
					},
					"address": {
						"type": "string"
					},
					"capacity": {
						"type": "integer"
					},
					"status": {
						"type": "string",
						"enum": ["ACTIVE", "INACTIVE", "MAINTENANCE"]
					},
					"operatingHours": {
						"type": "object",
						"properties": {
							"open": {
								"type": "string"
							},
							"close": {
								"type": "string"
							}
						}
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"SupportTicket": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"ticketNumber": {
						"type": "string"
					},
					"userId": {
						"type": "string",
						"format": "uuid"
					},
					"category": {
						"type": "string",
						"enum": ["ACCOUNT_ISSUE", "PAYMENT_PROBLEM", "TECHNICAL_SUPPORT", "BOOKING_ISSUE", "DRIVER_COMPLAINT", "FEATURE_REQUEST", "BUG_REPORT", "OTHER"]
					},
					"subject": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"priority": {
						"type": "string",
						"enum": ["LOW", "MEDIUM", "HIGH", "URGENT"]
					},
					"status": {
						"type": "string",
						"enum": ["OPEN", "IN_PROGRESS", "WAITING_FOR_USER", "RESOLVED", "CLOSED"]
					},
					"contactMethod": {
						"type": "string",
						"enum": ["EMAIL", "PHONE", "IN_APP"]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"PaginationMeta": {
				"type": "object",
				"properties": {
					"page": {
						"type": "integer"
					},
					"limit": {
						"type": "integer"
					},
					"total": {
						"type": "integer"
					},
					"totalPages": {
						"type": "integer"
					}
				}
			}
		}
	},
	"paths": {
...
		"/ranks": {
			"get": {
				"tags": ["Ranks"],
				"summary": "Get taxi ranks list",
				"parameters": [
					{
						"name": "status",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["ACTIVE", "INACTIVE", "MAINTENANCE"]
						}
					},
					{
						"name": "latitude",
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "longitude",
						"in": "query",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "radius",
						"in": "query",
						"schema": {
							"type": "number",
							"default": 10
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of taxi ranks",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"ranks": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Rank"
											}
										},
										"pagination": {
											"$ref": "#/components/schemas/PaginationMeta"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/support": {
			"get": {
				"tags": ["Support"],
				"summary": "Get support tickets",
				"parameters": [
					{
						"name": "status",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["OPEN", "IN_PROGRESS", "WAITING_FOR_USER", "RESOLVED", "CLOSED"]
						}
					},
					{
						"name": "category",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["ACCOUNT_ISSUE", "PAYMENT_PROBLEM", "TECHNICAL_SUPPORT", "BOOKING_ISSUE", "DRIVER_COMPLAINT", "FEATURE_REQUEST", "BUG_REPORT", "OTHER"]
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of support tickets",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"tickets": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/SupportTicket"
											}
										},
										"pagination": {
											"$ref": "#/components/schemas/PaginationMeta"
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Support"],
				"summary": "Create a new support ticket",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["category", "subject", "description"],
								"properties": {
									"category": {
										"type": "string",
										"enum": ["ACCOUNT_ISSUE", "PAYMENT_PROBLEM", "TECHNICAL_SUPPORT", "BOOKING_ISSUE", "DRIVER_COMPLAINT", "FEATURE_REQUEST", "BUG_REPORT", "OTHER"]
									},
									"subject": {
										"type": "string",
										"minLength": 5,
										"maxLength": 200
									},
									"description": {
										"type": "string",
										"minLength": 10,
										"maxLength": 5000
									},
									"priority": {
										"type": "string",
										"enum": ["LOW", "MEDIUM", "HIGH", "URGENT"],
										"default": "MEDIUM"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Support ticket created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SupportTicket"
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}